Add SunflowerApp composable.
