+++ app/src/main/java/com/google/samples/apps/sunflower/GardenActivity.kt
@@ -20,14 +20,7 @@ import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.WindowCompat
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
-import com.google.android.material.composethemeadapter.MdcTheme
-import com.google.samples.apps.sunflower.compose.home.Home
-import com.google.samples.apps.sunflower.compose.plantdetail.PlantDetails
+import com.google.samples.apps.sunflower.compose.SunflowerApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -40,27 +33,7 @@ class GardenActivity : AppCompatActivity() {
     WindowCompat.setDecorFitsSystemWindows(window, false)
 
     setContent {
-      MdcTheme {
-        val navController = rememberNavController()
-        NavHost(navController = navController, startDestination = "home") {
-          composable("home") {
-            Home(supportFragmentManager) { plantId ->
-              navController.navigate("plant/$plantId")
-            }
-          }
-          composable(
-            "plant/{plantId}",
-            arguments = listOf(navArgument("plantId") {
-              type = NavType.StringType
-            })
-          ) { backStackEntry ->
-            val plantId = backStackEntry.arguments?.getString("plantId")
-            PlantDetails(supportFragmentManager, plantId) {
-              navController.navigateUp()
-            }
-          }
-        }
-      }
+      SunflowerApp(fragmentManager = supportFragmentManager)
     }
   }
 }
