+++ app/src/main/java/com/google/samples/apps/sunflower/compose/SunflowerApp.kt
@@ -16,3 +16,38 @@
 
 package com.google.samples.apps.sunflower.compose
 
+import androidx.compose.runtime.Composable
+import androidx.fragment.app.FragmentManager
+import androidx.navigation.NavType
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navArgument
+import com.google.android.material.composethemeadapter.MdcTheme
+import com.google.samples.apps.sunflower.compose.home.Home
+import com.google.samples.apps.sunflower.compose.plantdetail.PlantDetails
+
+@Composable
+fun SunflowerApp(fragmentManager: FragmentManager) {
+  MdcTheme {
+    val navController = rememberNavController()
+    NavHost(navController = navController, startDestination = "home") {
+      composable("home") {
+        Home(fragmentManager) { plantId ->
+          navController.navigate("plant/$plantId")
+        }
+      }
+      composable(
+        "plant/{plantId}",
+        arguments = listOf(navArgument("plantId") {
+          type = NavType.StringType
+        })
+      ) { backStackEntry ->
+        val plantId = backStackEntry.arguments?.getString("plantId")
+        PlantDetails(fragmentManager, plantId) {
+          navController.navigateUp()
+        }
+      }
+    }
+  }
+}
